add async await feature

+++ server/server.js
@@ -77,24 +77,35 @@ app.get('/todos/:id', authenticate, (req, res) => {
 
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
 
   var id = req.params.id;
   if(!ObjectID.isValid(id)){
     return res.status(404).send();
   };
 
-ToDo.findOneAndRemove({
-  _id: id,
-  _creator: req.user._id
-}).then((todo) => {
+try {
+  var todo = await ToDo.findOneAndRemove({_id: id, _creator: req.user._id});
   if(!todo){
-    return res.status(404).send();
-  };
+      return res.status(404).send();
+    };
   res.send({todo});
-}).catch((e) => {
-  res.status(400).send();
-});
+} catch (e) {
+  res.status(404).send();
+};
+
+
+// ToDo.findOneAndRemove({
+//   _id: id,
+//   _creator: req.user._id
+// }).then((todo) => {
+//   if(!todo){
+//     return res.status(404).send();
+//   };
+//   res.send({todo});
+// }).catch((e) => {
+//   res.status(400).send();
+// });
 
 });
 
@@ -137,46 +148,72 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 
 });
 
-app.post('/users', (req, res) => {
+app.post('/users', async (req, res) => {
 
   var userParams = _.pick(req.body, ['email', 'password']);
   var userInput = new User(userParams);
+  try {
+    await userInput.save();
+    var token = await userInput.generateAuthToken();
+    res.header('x-auth', token).send(userInput);
+  } catch (e) {
+    res.status(400).send(e);
+  }
 
   //next two lines of code can be replaced with the commented line below
   //userInput.generateAuthToken().then((token) => {
 
-  userInput.save().then(() => {  //same as userInput.save().then((user); return user.generateAuthToken();
-    return userInput.generateAuthToken();
-  }).then((token) => {
-    res.header('x-auth', token).send(userInput);
-  }).catch((e) => {
-    res.status(400).send(e);
-  });
+  // userInput.save().then(() => {  //same as userInput.save().then((user); return user.generateAuthToken();
+  //   return userInput.generateAuthToken();
+  // }).then((token) => {
+  //   res.header('x-auth', token).send(userInput);
+  // }).catch((e) => {
+  //   res.status(400).send(e);
+  // });
 
 });
+
 
 app.get('/users/me', authenticate, (req, res) => {
   res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
 
-  User.findByCredentials(body.email, body.password).then((user) => {
-      return user.generateAuthToken().then((token) => {
-        res.header('x-auth', token).send(user);
-      });
-    }).catch((e) => {
+app.post('/users/login', async (req, res) => {
+
+  try {
+    var body = _.pick(req.body, ['email', 'password']);
+    var user = await User.findByCredentials(body.email, body.password);
+    var token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
     res.status(400).send();
-  });
+  }
+
+  // const User.findByCredentials(body.email, body.password).then((user) => {
+  //     return user.generateAuthToken().then((token) => {
+  //       res.header('x-auth', token).send(user);
+  //     });
+  //   }).catch((e) => {
+  //   res.status(400).send();
+  // });
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
-    res.status(200).send();
-  }).catch((e) => {
-    res.status(400).send();
-  })
+
+app.delete('/users/me/token', authenticate, async (req, res) => {
+try{
+  const removeToken = await req.user.removeToken(req.token);
+  res.status(200).send();
+} catch (e) {
+  res.status(400).send();
+}
+
+// replace below with async await
+  // req.user.removeToken(req.token).then(() => {
+  //   res.status(200).send();
+  // }).catch((e) => {
+  //   res.status(400).send();
+  // })
 });
 
 app.listen(port, () => {
